[Functional] Increase test_timeout for db migration tests
Patch [1] increased timeouts for test_walk_version functional tests
for MySQL backend to 300 seconds to avoid failures due to timeouts.

Unfortunately it looks that on nodes from some cloud providers used
in the gate and with number of migration scripts which we have in
Neutron those tests can take sometimes even around 400 seconds.

So lets increase this to 600 seconds to avoid such failures of
functional tests job.

[1] https://review.openstack.org/#/c/610003/

Change-Id: I9d129f0e90a072ec980aadabb2c6b812c08e1618
Closes-Bug: #1687027