Only store segmenthostmapping when enable segment plugin
This patch adds a check to determine if the 'segments' service plugin is
enabled. The segment-host mapping db table should only be saved and
updated to the db table if the users configure the 'segments' service
plugin in the config file. The data should be available only in a routed
network resource situation.

Change-Id: I65a42aa2129bef696906a18d82575461dc02ba21
Closes-Bug: #1799328