rally-jobs: Add FloatingIP workloads
As discussed on the neutron-performance meeting [1] floating IP
operations should be covered in the neutron gate rally job.
As I set out to write the missing tests I discovered that some of it
was already written by rally folks. For example see [2].
In neutron we only need to add them to the rally job definition.
Keep in mind the test code has been relocated since to the
rally-openstack repository. For example see [3].

[1] http://eavesdrop.openstack.org/meetings/neutron_performance/2018/neutron_performance.2018-10-22-16.02.log.html#l-143
[2] https://review.openstack.org/225223
[3] https://github.com/openstack/rally-openstack/blob/4e4bfc5/rally_openstack/scenarios/neutron/network.py

Change-Id: Icec7011cb293f1c92d968ef60efe7468ea4631fb