Reinitialize ovs firewall after ovs-vswitchd restart
When ovs-vswitchd process is restarted on host neutron-ovs-agent
can properly handle that and recover openflows and bridges config.
But when ovs firewall driver is used, it wasn't reinitialized
so there were missing some OF rules in br-int and connectivity
to/from instances was broken.

This patch adds reinitialization of firewall in such case so
all OF rules are properly reconfigured and connectivity is
restored.

Change-Id: I8d525bbe48216fef4da890582b8e198933a27833
Closes-Bug: 1805808
(cherry picked from commit 2ba9e95156d86bc67ecd8440d7343028e89cae3a)