remove context_manager from neutron
The context_manager is already in neutron-lib and being used by our
active consumers either directly or via the current shim we have
in neutron.db.api.context_manager that just references that in lib.
This patch removes the context_manager from neutron and updates the UTs
that mock it.

NeutronLibImpact

Change-Id: Ia5191a71b377e2862592a09da8dc3a5d415de9aa