Add ipxe6 dnsmasq tag to enable ipv6 dhcp client matching
DHCPv6 has different mode of operation and the request class 175
option no longer works for iPXE to allow configurations to be set
that delineate the address requestor as an iPXE client or a ROM
client seeking a bootloader over via DHCPv6.

In order to support iPXE, ironic needs to be able to match on to
the user class that is transmitted in the DHCPv6 requests.
From there, dnsmasq can choose the correct reply and a node can
boot up to be deployed by ironic.

Change-Id: I6328297deaabdc94b6356930dd978b02f432799c
Story: 2004502
Task: 28243