Pass elevated admin context when fetching snat router for FIP to unbound port
If a floating IP is assigned to an unbound port, it is scheduled to the SNAT router.
This fails when the router is in a different tenant and user is not an admin.
This was previously working, but broken with the changes for enabling
centralized FIP/unbound port.

A common usecase for this when not using a true VIP is a port and Floating IP may be
precreated as a pair, and then attached to a VM.
It works otherwise if the port is already bound to a VM.

Change-Id: I0ee5344216fdff0a9e3bd6fa165f8d0aa736ac85
Closes-Bug: #1802006