Use oslo_log instead of alembic.util in functional tests
Instead of using alembic.utils logging functions which
prints output to stdout, lets use in functional tests
oslo_log logger to log all such commands.

This is needed to limit amount of output on stdout during
functional tests running on python3.5.

It was previously switched to use oslo_log in this migration
scripts always by [1] but because of issue in OSA we had to
revert this patch in [2].
So now let's try to change it only in functional tests.

[1] https://review.openstack.org/#/c/625555/
[2] https://review.openstack.org/#/c/626565/

Change-Id: I88fde6d107a1f1b5c48c5c89cd2e1c07fd12c755