Revert "Remove -u root as mysql is executed with root user"
This commit was approved just to keep the consistency with
other OpenStack projects, but most proposed changes related to it
have been rejected so far [1].
The main reason is because it is clearer to specify options explicitly
and we should not assume some configuration.

I totally agree these opinions and see no reason that neutron
contineus to use different command line options for mysql.
Consistency in the installation guide will bring better user experience.

[1] https://review.openstack.org/#/q/file:%255Edoc/source/install/.*+message:mysql

This reverts commit 59d178c9de6028016053f2a6958372e14ded386a.

Related-Bug: #1785025
Change-Id: I699d03de9e26d5784e6bd5e7d4fa219de6209207