Define types for C calls in netlink_lib
Previously this was not done, which meant all arguments were
assumed to be ints. As long as we didn't get any large pointer
addresses this worked fine, but for some reason the addition of
threading to oslo.privsep triggered larger addresses that were then
truncated. This caused segfaults in the underlying C library because
we were passing it invalid pointers.

This change sets argument and return types for all of the calls
that are used in the module.

Change-Id: I9dbbb01eac8cf7cded7794f8ba69797d6357eeec
Closes-Bug: 1810518