Change common agent code to log message after failure
If an agent using the common agent code fails to send
its report_state to the server, it logs an exception:

   Failed reporting state!: MessagingTimeout: Timed out...

If it then tries a second time and succeeds it just
goes on happily. It would be nice if it logged that
it had success on the subsequent attempt so someone
looking at the logs know it recovered.

Change-Id: Ie94e909fba07e5eb91b57160947edc9b86428fe5