Run revision bump operations en masse
Refactored RevisionPlugin to operate upon sets of objects at
once.   In the case of "related" objects, the bump operation
no longer makes use of compare-and-swap and instead updates
version numbers directly without testing for their previous value.
This removes the issue of StaleDataErrors being
prevalent within the related object update phase, given the assumption
that the compare-and-swap logic is only desireable for the primary object
being updated.

Change-Id: I2fef298041c59a03dfd06912764973995b80690c