Remove redundant event listening logic from neutron/db/api
In I3e702b99fd5084e8090f93c384aa1f704edceaff, the event listening
features which handle expiration of relationships was "rehomed"
to neutron-lib.   In fact it was duplicated, so that these events
run twice.    This patch removes the events from neutron as it is
a very bad idea for this much complexity to be running twice.

Change-Id: I8d0321487e4b406ead7a237156ef2f221fb512b9