Add new test decorator skip_if_timeout
In some cases our db migration tests which run on MySQL are
failing with timeout and it happens due to slow VMs on which
job is running.
Sometimes it may also happen that timeout exception is raised
in the middle of some sqlalchemy operations and
sqlalchemy.InterfaceError is raised as last one.
Details about this exception can be found in [1].

To avoid many rechecks because of this reason this patch
introduces new decorator which is very similar to "unstable_test"
but will skip test only if one of exceptions mentioned above will
be raised.
In all other cases it will fail test.

That should be a bit more safe for us because we will not miss
some other failures raised in those tests and will avoid rechecks
because of this "well-known" reason described in related bug.

[1] http://sqlalche.me/e/rvf5

Conflicts:
    neutron/tests/functional/db/test_migrations.py
    neutron/tests/base.py

Change-Id: Ie291fda7d23a696aaa1160d126a3cf72b08c522f
Related-Bug: #1687027
(cherry picked from commit c0fec676723649a0516cf3d4af0dccc0fe832095)
(cherry picked from commit e6f22ce81c0a1130d45d290b185b736501e6dd1e)