DVR-HA: Unbinding a HA router from agent does not clear HA interface
Removing an active or a standby HA router from an agent that has a
valid DVR serviceable port (such as DHCP), does not remove the
HA interface associated with the Router in the SNAT namespace.

When we try to add the HA router back to the agent, then it
adds more than one HA interface to the SNAT Namespace causing
more problems and we sometimes also see multiple active routers.

This bug might have been introduced by this patch [1].

Fix the problem by just adding the router namespaces without HA
interfaces when there is no HA and re-insert the HA interfaces
when HA router is bound to the agent into the namespace.

[1] https://review.openstack.org/#/c/522362/
Closes-Bug: #1816698

Change-Id: Ie625abcb73f8185bb2bee06dcd26a01d8af0b0d1