[Functional tests] Change way how conntrack entries are checked
In netlink_lib functional tests module there are listed conntrack
entries and those entries are assert to some expected list.
It may happen that sometimes some additional entries from other
tests will also be in the list and that cause failures of
netlink_lib tests.

So this patch changes way how those assertions are done. For now
it will check if each of expected entries is in entries list and
in case of delete entries tests, it will also check if any of
deleted entries isn't actually in list.

Change-Id: I30c18f141a8356b060902e6493ba0657b21619ad
Closes-Bug: #1817295